dwarf2out.c (DWARF_CIE_DATA_ALIGNMENT): Adjust, now that UNITS_PER_WORD is unsigned.
	* dwarf2out.c (DWARF_CIE_DATA_ALIGNMENT): Adjust, now that
	UNITS_PER_WORD is unsigned.

From-SVN: r33200

+++ gcc/dwarf2out.c
@@ -164,9 +164,9 @@ dw_fde_node;
 
 /* Offsets recorded in opcodes are a multiple of this alignment factor.  */
 #ifdef STACK_GROWS_DOWNWARD
-#define DWARF_CIE_DATA_ALIGNMENT (-UNITS_PER_WORD)
+#define DWARF_CIE_DATA_ALIGNMENT (-((int) UNITS_PER_WORD))
 #else
-#define DWARF_CIE_DATA_ALIGNMENT UNITS_PER_WORD
+#define DWARF_CIE_DATA_ALIGNMENT ((int) UNITS_PER_WORD)
 #endif
 
 /* A pointer to the base of a table that contains frame description
